refactor: Refactor todo model to add filtering of items by user id.
- Modified `GetOneTodo()` to filter by both todo id and user id.
- Modified `GetAllTodos()` to filter by user id.
- Modified `InsertOneTodo()` to include a user id in the query.
- Modified `UpdateOneTodo()` to filter by both todo id and user id.
- Modified `DeleteOneTodo()` to filter by both todo id and user id.

+++ models/todo.go
@@ -7,28 +7,29 @@ type Todo struct {
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	Done        bool   `json:"done"`
+	UserID      int64  `json:"user_id"`
 }
 
-func GetOneTodo(id int64) (todo Todo, err error) {
+func GetOneTodo(id int64, userID int64) (todo Todo, err error) {
 	db, err := db.Open()
 	if err != nil {
 		return
 	}
 
-	row := db.QueryRow("SELECT * FROM todos WHERE id=$1", id)
-	err = row.Scan(&todo.Id, &todo.Title, &todo.Description, &todo.Done)
+	row := db.QueryRow("SELECT * FROM todos WHERE id=$1 AND user_id=$2", id, userID)
+	err = row.Scan(&todo.Id, &todo.Title, &todo.Description, &todo.Done, &todo.UserID)
 
 	db.Close()
 	return
 }
 
-func GetAllTodos() (todos []Todo, err error) {
+func GetAllTodos(userID int64) (todos []Todo, err error) {
 	db, err := db.Open()
 	if err != nil {
 		return
 	}
 
-	rows, err := db.Query("SELECT * FROM todos")
+	rows, err := db.Query("SELECT * FROM todos WHERE user_id=$1", userID)
 	if err != nil {
 		return
 	}
@@ -36,7 +37,7 @@ func GetAllTodos() (todos []Todo, err error) {
 	for rows.Next() {
 		var todo Todo
 
-		err = rows.Scan(&todo.Id, &todo.Title, &todo.Description, &todo.Done)
+		err = rows.Scan(&todo.Id, &todo.Title, &todo.Description, &todo.Done, &todo.UserID)
 		if err != nil {
 			continue
 		}
@@ -48,26 +49,26 @@ func GetAllTodos() (todos []Todo, err error) {
 	return
 }
 
-func InsertOneTodo(todo Todo) (id int64, err error) {
+func InsertOneTodo(todo Todo, userID int64) (id int64, err error) {
 	db, err := db.Open()
 	if err != nil {
 		return
 	}
 
-	sql := "INSERT INTO todos (title, description, done) VALUES ($1, $2, $3) RETURNING id"
-	err = db.QueryRow(sql, todo.Title, todo.Description, todo.Done).Scan(&id)
+	sql := "INSERT INTO todos (title, description, done, user_id) VALUES ($1, $2, $3, $4) RETURNING id"
+	err = db.QueryRow(sql, todo.Title, todo.Description, todo.Done, userID).Scan(&id)
 
 	db.Close()
 	return
 }
 
-func UpdateOneTodo(id int64, todo Todo) (int64, error) {
+func UpdateOneTodo(id int64, todo Todo, userID int64) (int64, error) {
 	db, err := db.Open()
 	if err != nil {
 		return 0, err
 	}
 
-	res, err := db.Exec(`UPDATE todos SET title=$1, description=$2, done=$3 WHERE id=$4`, todo.Title, todo.Description, todo.Done, id)
+	res, err := db.Exec(`UPDATE todos SET title=$1, description=$2, done=$3 WHERE id=$4 AND user_id=$5`, todo.Title, todo.Description, todo.Done, id, userID)
 	if err != nil {
 		return 0, err
 	}
@@ -76,13 +77,13 @@ func UpdateOneTodo(id int64, todo Todo) (int64, error) {
 	return res.RowsAffected()
 }
 
-func DeleteOneTodo(id int64) (int64, error) {
+func DeleteOneTodo(id int64, userID int64) (int64, error) {
 	db, err := db.Open()
 	if err != nil {
 		return 0, err
 	}
 
-	res, err := db.Exec(`DELETE FROM todos WHERE id=$1`, id)
+	res, err := db.Exec(`DELETE FROM todos WHERE id=$1 AND user_id=$2`, id, userID)
 	if err != nil {
 		return 0, err
 	}
